Trait pwbox::Cipher

pub trait Cipher: 'static {
    const KEY_LEN: usize;
    const NONCE_LEN: usize;
    const MAC_LEN: usize;

    fn seal(message: &[u8], nonce: &[u8], key: &[u8]) -> CipherOutput;

    fn open(
        output: &mut [u8], 
        encrypted: &CipherOutput, 
        nonce: &[u8], 
        key: &[u8]
    ) -> Result<(), MacMismatch>;
}

Authenticated symmetric cipher.

Associated Constants

const KEY_LEN: usize

Byte size of a key.

const NONCE_LEN: usize

Byte size of a nonce (aka initialization vector, or IV).

const MAC_LEN: usize

Byte size of a message authentication code (MAC).

Loading content...
Required methods

fn seal(message: &[u8], nonce: &[u8], key: &[u8]) -> CipherOutput

Encrypts message with the provided key and nonce.

Safety

When used within PwBox, key and nonce are guaranteed to have correct sizes.

fn open(
output: &mut [u8],
encrypted: &CipherOutput,
nonce: &[u8],
key: &[u8]
) -> Result<(), MacMismatch>

Decrypts encrypted message with the provided key and nonce and stores the result into output. If the MAC does not verify, returns an error.

Safety

When used within PwBox, key, nonce, encrypted.mac and output are guaranteed to have correct sizes.

Loading content...